Migrated 'number_from_name' test to its own test module
